chore: Make Lighthouse happier
