+++ .eslintrc.json
@@ -8,7 +8,11 @@
     "globals": {
         "__webpack_public_path__": "writable"
     },
-    "extends": ["eslint:recommended", "plugin:prettier/recommended","plugin:jest/recommended"],
+    "extends": [
+        "eslint:recommended",
+        "plugin:prettier/recommended",
+        "plugin:jest/recommended"
+    ],
     "parserOptions": {
         "ecmaVersion": 12,
         "sourceType": "module"
