+++ .github/workflows/build.yml
@@ -1,24 +1,26 @@
 name: Build
 on:
-  pull_request:
-    branches: [ master ]
+    pull_request:
+        branches: [master]
 
 jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2.3.1 
+    build:
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v2.3.1
 
-      - name: Install dependencies
-        run: |
-          yarn --silent
+            - name: Install dependencies
+              run: |
+                  yarn --silent
+                  docker pull emscripten/emsdk
 
-      - name: Test and Lint
-        run: |
-          yarn lint
-          yarn test:ci
+            - name: Test and Lint
+              run: |
+                  yarn lint
+                  yarn wasm:production
+                  yarn test:ci
 
-      - name: Production build
-        run: |
-          yarn build
+            - name: Production build
+              run: |
+                  yarn build
