+++ .github/workflows/deploy.yml
@@ -1,34 +1,36 @@
 name: Build and Deploy
 on:
-  push:
-    branches: [ master ]
+    push:
+        branches: [master]
 
 jobs:
-  build-and-deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2.3.1 
-        with:
-          persist-credentials: false
+    build-and-deploy:
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v2.3.1
+              with:
+                  persist-credentials: false
 
-      - name: Install dependencies
-        run: |
-          yarn --silent
+            - name: Install dependencies
+              run: |
+                  yarn --silent
+                  docker pull emscripten/emsdk
 
-      - name: Test and Lint
-        run: |
-          yarn lint
-          yarn test:ci
+            - name: Test and Lint
+              run: |
+                  yarn lint
+                  yarn wasm:production
+                  yarn test:ci
 
-      - name: Production build
-        run: |
-          yarn build
+            - name: Production build
+              run: |
+                  yarn build
 
-      - name: Deploy
-        uses: JamesIves/github-pages-deploy-action@3.7.1
-        with:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          BRANCH: gh-pages
-          FOLDER: dist
-          CLEAN: true
+            - name: Deploy
+              uses: JamesIves/github-pages-deploy-action@3.7.1
+              with:
+                  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+                  BRANCH: gh-pages
+                  FOLDER: dist
+                  CLEAN: true
