+++ .prettierrc.json
@@ -3,4 +3,4 @@
     "tabWidth": 4,
     "semi": true,
     "singleQuote": false
-  }
+}
