+++ babel.config.js
@@ -1,0 +1,4 @@
+// babel.config.js
+module.exports = {
+    presets: [["@babel/preset-env", { targets: { node: "current" } }]],
+};
