+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    transform: {
+        "\\.js$": "babel-jest",
+        "\\.wasm$": ["jest-file-loader", "babel-jest"],
+    },
+};
