+++ src/index.ejs
@@ -1,0 +1,51 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <head>
+    <meta charset="utf-8">
+    <title><%= htmlWebpackPlugin.options.title %></title>
+    <style>
+        body {
+            background-color: #f5f2f0;
+        }
+        #spinner,
+        #spinner:after {
+            border-radius: 50%;
+            width: 50vmin;
+            height: 50vmin;
+        }
+        #spinner {
+            font-size: 10px;
+            position: fixed;
+            top: 50%;
+            left: 50%;
+            text-indent: -9999em;
+            pointer-events: none;
+            border-top: 1.1em solid rgba(51,51,51, 0.2);
+            border-right: 1.1em solid rgba(51,51,51, 0.2);
+            border-bottom: 1.1em solid rgba(51,51,51, 0.2);
+            border-left: 1.1em solid #333333;
+            transform: translateZ(100);
+            animation: load8 1.1s infinite linear;
+        }
+        @-webkit-keyframes load8 {
+            0% {
+                transform: translate(-50%, -50%) rotate(0deg);
+            }
+            100% {
+                transform: translate(-50%, -50%)  rotate(360deg);
+            }
+        }
+        @keyframes load8 {
+            0% {
+                transform: translate(-50%, -50%) rotate(0deg);
+            }
+            100% {
+                transform: translate(-50%, -50%) rotate(360deg);
+            }
+        }
+    </style>
+  </head>
+  <body>
+    <div id="spinner">Loading...</div>
+  </body>
+</html>
