+++ src/runtime/rubiks.h
@@ -1,0 +1,3 @@
+#define OPERATION_COUNT 54
+#define OPT_LOOKUP_LENGTH 3
+#define OPT_LOOKUP_SIZE (OPERATION_COUNT * OPERATION_COUNT * OPERATION_COUNT)
