+++ src/styles/neishauben.css
@@ -11,7 +11,7 @@ canvas {
     right: 0;
     display: flex;
     justify-content: center;
-    align-items: center;   
+    align-items: center;
     flex-wrap: wrap;
 }
 
@@ -24,7 +24,7 @@ canvas {
     flex-direction: column;
     display: flex;
     justify-content: center;
-    align-items: center;   
+    align-items: center;
     flex-wrap: wrap;
 }
 
@@ -34,10 +34,10 @@ canvas {
     display: flex;
     flex-wrap: wrap;
     justify-content: center;
-    align-items: center;   
+    align-items: center;
     background-color: rgba(245, 242, 240, 0.7);
     border-radius: 10px;
-    padding: 5px;   
+    padding: 5px;
 }
 
 #algorithm a {
@@ -50,7 +50,8 @@ a.algorithm-link {
     text-decoration: inherit;
 }
 
-a.algorithm-link[href]:hover, a.algorithm-current {
+a.algorithm-link[href]:hover,
+a.algorithm-current {
     background-color: #3484d5;
     box-sizing: content-box;
     color: white;
@@ -75,8 +76,8 @@ a.algorithm-done {
     bottom: 0;
 }
 
-#top-bar >*,
-#bottom-bar >* {
+#top-bar > *,
+#bottom-bar > * {
     margin: 5px;
 }
