+++ webpack.config.js
@@ -1,8 +1,9 @@
 const path = require("path");
+const { IgnorePlugin } = require("webpack");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
-const FaviconsWebpackPlugin = require('favicons-webpack-plugin')
-const WorkboxPlugin = require('workbox-webpack-plugin');
+const FaviconsWebpackPlugin = require("favicons-webpack-plugin");
+const WorkboxPlugin = require("workbox-webpack-plugin");
 
 module.exports = {
     context: path.resolve(__dirname, "."),
@@ -10,54 +11,51 @@ module.exports = {
     output: {
         path: path.resolve(__dirname, "dist"),
         filename: "[name].[contenthash:8].js",
-        publicPath: "/neishauben/"
+        publicPath: "/neishauben/",
     },
     devtool: "source-map",
     plugins: [
+        new IgnorePlugin(/fs/),
         new CleanWebpackPlugin(),
         new FaviconsWebpackPlugin({
-            logo: './static/logo.png',
-            mode: 'webapp',
-            devMode: 'webapp',
+            logo: "./static/logo.png",
+            mode: "webapp",
+            devMode: "webapp",
             favicons: {
-              appName: 'Neishauben',
-              appDescription: 'Rubik\'s Cube Simulator',
-              developerName: 'Tibor Durica Potpara',
-              developerURL: 'https://ojdip.net',
-              background: '#f5f2f0',
-              theme_color: '#3484d5',
-              scope: "/neishauben/",
-              start_url: "/neishauben/",
-              icons: {
-                coast: false,
-                yandex: false
-              }
-            }
-          }),          
+                appName: "Neishauben",
+                appDescription: "Rubik's Cube Simulator",
+                developerName: "Tibor Durica Potpara",
+                developerURL: "https://ojdip.net",
+                background: "#f5f2f0",
+                theme_color: "#3484d5",
+                scope: "/neishauben/",
+                start_url: "/neishauben/",
+                icons: {
+                    coast: false,
+                    yandex: false,
+                },
+            },
+        }),
+        new HtmlWebpackPlugin({
+            title: "Neishauben",
+            meta: {
+                description: "Rubik's Cube Simulator",
+                viewport:
+                    "width=device-width, initial-scale=1, shrink-to-fit=no, maximum-scale=1.0, user-scalable=no",
+            },
+        }),
         new WorkboxPlugin.GenerateSW({
             // these options encourage the ServiceWorkers to get in there fast
             // and not allow any straggling "old" SWs to hang around
             clientsClaim: true,
             skipWaiting: true,
-        }),          
-        new HtmlWebpackPlugin({
-            title: "Neishauben",
-            meta: {
-                viewport:
-                    "width=device-width, initial-scale=1, shrink-to-fit=no, maximum-scale=1.0, user-scalable=no",
-            },
+            directoryIndex: "index.html",
         }),
     ],
     optimization: {
         runtimeChunk: "single",
         splitChunks: {
-            cacheGroups: {
-                vendor: {
-                    test: /[\\/]node_modules[\\/]/,
-                    name: "vendors",
-                    chunks: "all",
-                },
-            },
+            chunks: "all",
         },
     },
     module: {
@@ -67,12 +65,16 @@ module.exports = {
                 type: "javascript/auto",
                 loader: "file-loader",
             },
-            {test: /\.(jpe?g|png|gif|svg)$/i, loader: "file-loader"},
+            {
+                test: /\.worker\.js$/,
+                use: { loader: "worker-loader" },
+            },
+            { test: /\.(jpe?g|png|gif|svg)$/i, loader: "file-loader" },
             { test: /\.css$/, use: ["style-loader", "css-loader"] },
         ],
     },
     devServer: {
         contentBase: path.join(__dirname, "dist"),
-        publicPath: "/neishauben/"
+        publicPath: "/neishauben/",
     },
 };
